moved KeyHandler.kt to helpers
